[4145] Remove editing context event handler execution from the EditingContextEventProcessor
Bug: https://github.com/eclipse-sirius/sirius-web/issues/4145
