+++ packages/core/backend/sirius-components-collaborative/src/main/java/org/eclipse/sirius/components/collaborative/api/IEditingContextExecutorFactory.java
@@ -1,0 +1,23 @@
+/*******************************************************************************
+ * Copyright (c) 2024 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.sirius.components.collaborative.api;
+
+/**
+ * The factory used to create an editing context executor.
+ *
+ * @author gcoutable
+ */
+public interface IEditingContextExecutorFactory {
+
+    IEditingContextExecutor createEditingContextExecutor();
+}
