+++ packages/core/backend/sirius-components-collaborative/src/main/java/org/eclipse/sirius/components/collaborative/editingcontext/EditingContextEventProcessor.java
@@ -12,7 +12,6 @@
  *******************************************************************************/
 package org.eclipse.sirius.components.collaborative.editingcontext;
 
-import java.time.Duration;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -21,15 +20,12 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
-import java.util.logging.Level;
 
 import org.eclipse.sirius.components.collaborative.api.ChangeDescription;
 import org.eclipse.sirius.components.collaborative.api.ChangeKind;
 import org.eclipse.sirius.components.collaborative.api.IChangeDescriptionListener;
-import org.eclipse.sirius.components.collaborative.api.IEditingContextEventHandler;
 import org.eclipse.sirius.components.collaborative.api.IEditingContextEventProcessor;
-import org.eclipse.sirius.components.collaborative.api.IInputPostProcessor;
-import org.eclipse.sirius.components.collaborative.api.IInputPreProcessor;
+import org.eclipse.sirius.components.collaborative.api.IEditingContextExecutor;
 import org.eclipse.sirius.components.collaborative.api.IRepresentationEventProcessor;
 import org.eclipse.sirius.components.collaborative.api.IRepresentationEventProcessorRegistry;
 import org.eclipse.sirius.components.collaborative.api.Monitoring;
@@ -38,20 +34,15 @@ import org.eclipse.sirius.components.core.api.ErrorPayload;
 import org.eclipse.sirius.components.core.api.IEditingContext;
 import org.eclipse.sirius.components.core.api.IInput;
 import org.eclipse.sirius.components.core.api.IPayload;
-import org.eclipse.sirius.components.core.api.IRepresentationInput;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.micrometer.core.instrument.MeterRegistry;
-import io.micrometer.core.instrument.Timer;
 import reactor.core.Disposable;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.core.publisher.SignalType;
 import reactor.core.publisher.Sinks;
 import reactor.core.publisher.Sinks.EmitResult;
 import reactor.core.publisher.Sinks.Many;
-import reactor.core.publisher.Sinks.One;
 
 /**
  * Handles all the inputs which concern a particular editing context one at a time, in order of arrival, and in a
@@ -62,15 +53,7 @@ import reactor.core.publisher.Sinks.One;
  */
 public class EditingContextEventProcessor implements IEditingContextEventProcessor {
 
-    public static final String INPUT = "INPUT";
-
     private final Logger logger = LoggerFactory.getLogger(EditingContextEventProcessor.class);
-
-    private final ICollaborativeMessageService messageService;
-
-    private final IEditingContext editingContext;
-
-    private final List<IEditingContextEventHandler> editingContextEventHandlers;
 
     private final IChangeDescriptionListener changeDescriptionListener;
 
@@ -82,17 +65,13 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
 
     private final Many<ChangeDescription> changeDescriptionSink = Sinks.many().unicast().onBackpressureBuffer();
 
-    private final ExecutorService executorService;
-
     private final Disposable changeDescriptionDisposable;
 
-    private final List<IInputPreProcessor> inputPreProcessors;
+    private final IEditingContextExecutor editingContextExecutor;
 
-    private final List<IInputPostProcessor> inputPostProcessors;
+    private final IEditingContext editingContext;
 
-    private final MeterRegistry meterRegistry;
-
-    public EditingContextEventProcessor(IRepresentationEventProcessorRegistry representationEventProcessorRegistry, EditingContextEventProcessorParameters parameters) {
+    public EditingContextEventProcessor(IRepresentationEventProcessorRegistry representationEventProcessorRegistry, IEditingContextExecutor editingContextExecutor, EditingContextEventProcessorParameters parameters) {
         this.representationEventProcessorRegistry = Objects.requireNonNull(representationEventProcessorRegistry);
         this.messageService = parameters.messageService();
         this.editingContext = parameters.editingContext();
@@ -101,6 +80,7 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
         this.executorService = parameters.executorServiceProvider().getExecutorService(this.editingContext);
         this.inputPreProcessors = parameters.inputPreProcessors();
         this.inputPostProcessors = parameters.inputPostProcessors();
+        this.editingContextExecutor = Objects.requireNonNull(editingContextExecutor);
         this.changeDescriptionDisposable = this.setupChangeDescriptionSinkConsumer();
         this.meterRegistry = parameters.meterRegistry();
     }
@@ -128,86 +108,12 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
 
     @Override
     public Mono<IPayload> handle(IInput input) {
-        Timer.Sample handleTimer = Timer.start(this.meterRegistry);
-        if (this.executorService.isShutdown()) {
-            this.logger.warn("Handler for editing context {} is shutdown", this.editingContext.getId());
-            handleTimer.stop(this.meterRegistry.timer(Monitoring.EVENT_HANDLER, INPUT, input.getClass().getSimpleName(),
-                    "inputId", input.id().toString()));
-            return Mono.empty();
-        }
-
-        this.logger.trace(input.toString());
-
-        One<IPayload> payloadSink = Sinks.one();
-        Future<?> future = this.executorService.submit(() -> this.doHandle(payloadSink, input));
-        try {
-            // Block until the event has been processed
-            future.get();
-        } catch (InterruptedException | ExecutionException exception) {
-            this.logger.warn(exception.getMessage(), exception);
-        }
-        handleTimer.stop(this.meterRegistry.timer(Monitoring.TIMER_PROCESSING_INPUT, "input", input.getClass().getSimpleName(),
-                "inputId", input.id().toString()));
-
-        var timeoutFallback = Mono.just(new ErrorPayload(input.id(), this.messageService.timeout()))
-                .doOnSuccess(payload -> this.logger.warn("Timeout fallback for the input {}", input));
-        return payloadSink.asMono()
-                .log(this.getClass().getName(), Level.FINEST, SignalType.ON_NEXT, SignalType.ON_ERROR)
-                .timeout(Duration.ofSeconds(5), timeoutFallback)
-                .doOnError(throwable -> this.logger.warn(throwable.getMessage(), throwable));
-    }
-
-    /**
-     * Finds the proper event handler to perform the task matching the given input event.
-     *
-     * @param payloadSink
-     *         The sink to publish payload
-     * @param input
-     *         The input event
-
-     */
-    private void doHandle(One<IPayload> payloadSink, IInput input) {
-        this.logger.trace("Input received: {}", input);
-
-        AtomicReference<IInput> inputAfterPreProcessing = new AtomicReference<>(input);
-        this.inputPreProcessors.forEach(preProcessor -> inputAfterPreProcessing.set(preProcessor.preProcess(this.editingContext, inputAfterPreProcessing.get(), this.changeDescriptionSink)));
-
-        if (inputAfterPreProcessing.get() instanceof IRepresentationInput representationInput) {
-            this.handleRepresentationInput(payloadSink, representationInput);
-        } else {
-            this.handleInput(payloadSink, inputAfterPreProcessing.get());
-        }
-
-        this.inputPostProcessors.forEach(postProcessor -> postProcessor.postProcess(this.editingContext, inputAfterPreProcessing.get(), this.changeDescriptionSink));
-    }
-
-    private void handleInput(One<IPayload> payloadSink, IInput input) {
-        Optional<IEditingContextEventHandler> optionalEditingContextEventHandler = this.editingContextEventHandlers.stream()
-                .filter(handler -> handler.canHandle(this.editingContext, input))
-                .findFirst();
-
-        if (optionalEditingContextEventHandler.isPresent()) {
-            IEditingContextEventHandler editingContextEventHandler = optionalEditingContextEventHandler.get();
-            editingContextEventHandler.handle(payloadSink, this.changeDescriptionSink, this.editingContext, input);
-        } else {
-            this.logger.warn("No handler found for event: {}", input);
-        }
-    }
-
-    private void handleRepresentationInput(One<IPayload> payloadSink, IRepresentationInput representationInput) {
-        Optional<IRepresentationEventProcessor> optionalRepresentationEventProcessor = this.acquireRepresentationEventProcessor(representationInput.representationId(), representationInput);
-
-        if (optionalRepresentationEventProcessor.isPresent()) {
-            IRepresentationEventProcessor representationEventProcessor = optionalRepresentationEventProcessor.get();
-            representationEventProcessor.handle(payloadSink, this.changeDescriptionSink, representationInput);
-        } else {
-            this.logger.warn("No representation event processor found for event: {}", representationInput);
-        }
+        return this.editingContextExecutor.handle(input);
     }
 
     @Override
     public Optional<IRepresentationEventProcessor> acquireRepresentationEventProcessor(String representationId, IInput input) {
-        return this.representationEventProcessorRegistry.getOrCreateRepresentationEventProcessor(representationId, this.editingContext, this.canBeDisposedSink, this.executorService);
+        return this.representationEventProcessorRegistry.getOrCreateRepresentationEventProcessor(representationId, this.editingContext, this.canBeDisposedSink, this.editingContextExecutor);
     }
 
     @Override
@@ -236,7 +142,7 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
         }
         this.changeDescriptionDisposable.dispose();
 
-        this.executorService.shutdown();
+        this.editingContextExecutor.dispose();
 
         this.representationEventProcessorRegistry.dispose();
